YYC-265 (A1): split config.rs into mod.rs + tests.rs
`src/config.rs` had grown to 2257 lines, with ~855 lines of tests
piled inside a single inline `mod tests {}` block. Split the file
into:

- `src/config/mod.rs` (1407 lines) — all configuration types,
  default fns, migration, atomic_write, prune helpers.
- `src/config/tests.rs` (855 lines) — the tests, declared via
  `#[cfg(test)] mod tests;` from `mod.rs`.

This is the smallest meaningful split: it pulls the test bulk
out of the main file (every navigation through this module no
longer scrolls past 800 lines of tests) without changing any
public API or moving any production code between files.

Further decomposition (provider / gateway / scheduler /
embeddings into their own submodules) is deferred — each cut
needs careful re-export plumbing for the dozens of `use
crate::config::Foo` sites elsewhere. That can land per-section
in follow-up PRs once the test-extraction pattern is settled.

`cargo test --features gateway` — 878/878 still pass.
